requirements.txt
APScheduler==3.10.4
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
defusedxml==0.7.1
elastic-transport==8.15.1
elasticsearch==8.15.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
langdetect==1.0.9
lxml==5.3.0
MarkupSafe==3.0.1
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
selenium==4.25.0
service-identity==24.1.0
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tika==2.6.0
tldextract==5.1.2
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
w3lib==2.2.1
watchdog==5.0.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.interface==7.1.0
PyMuPDF==1.24.10
python-docx==1.1.2
Aspose.Slides==24.10.0
pydrawing==0.1.9
moviepy==1.0.3
openpyxl==3.1.5